Creact a reusable component for ResponseCodeAbortButton to prevent duplicate code
