+++ src/components/ResetPassword/EmailLinkSent.tsx
@@ -1,17 +1,12 @@
 import { verifyEmailLink } from "apis/eduidResetPassword";
-import EduIDButton from "components/Common/EduIDButton";
+import { ResponseCodeButtons } from "components/Common/ResponseCodeAbortButton";
 import { ResponseCodeForm, ResponseCodeValues } from "components/Login/ResponseCodeForm";
 import { useAppDispatch, useAppSelector } from "eduid-hooks";
 import React, { useContext } from "react";
-import { FormRenderProps } from "react-final-form";
 import { FormattedMessage } from "react-intl";
 import { clearNotifications } from "slices/Notifications";
 import resetPasswordSlice from "slices/ResetPassword";
 import { ResetPasswordGlobalStateContext } from "./ResetPasswordGlobalState";
-
-interface ResponseCodeButtonsProps {
-  formProps?: FormRenderProps<ResponseCodeValues>;
-}
 
 export function EmailLinkSent(): JSX.Element | null {
   const dispatch = useAppDispatch();
@@ -46,47 +41,6 @@ export function EmailLinkSent(): JSX.Element | null {
       dispatch(resetPasswordSlice.actions.resetEmailStatus());
       resetPasswordContext.resetPasswordService.send({ type: "GO_BACK" });
     }
-  }
-
-  function ResponseCodeButtons(props: ResponseCodeButtonsProps) {
-    if (!props.formProps) {
-      return null;
-    }
-
-    // 'convert' from FormRenderProps to a simple "disabled" boolean
-    return (
-      <ResponseCodeAbortButton
-        disabled={props.formProps.submitting}
-        invalid={props.formProps.invalid}
-        submit={props.formProps.form.submit}
-      />
-    );
-  }
-
-  function ResponseCodeAbortButton(props: { disabled: boolean; invalid: boolean; submit: () => void }) {
-    // abort button usable from both ResponseCodeButtons and when isExpired below
-    return (
-      <div className="buttons">
-        <EduIDButton
-          type="button"
-          buttonstyle="secondary"
-          onClick={handleAbortButtonOnClick}
-          id="response-code-abort-button"
-          disabled={props.disabled}
-        >
-          <FormattedMessage defaultMessage="Cancel" description="Short code form" />
-        </EduIDButton>
-        <EduIDButton
-          type="submit"
-          buttonstyle="primary"
-          onClick={props.submit}
-          id="response-code-ok-button"
-          disabled={props.invalid}
-        >
-          <FormattedMessage defaultMessage="Ok" description="Short code form Ok button" />
-        </EduIDButton>
-      </div>
-    );
   }
 
   if (!response) {
@@ -124,7 +78,7 @@ export function EmailLinkSent(): JSX.Element | null {
       </p>
       <div className="enter-code">
         <ResponseCodeForm inputsDisabled={false} handleSubmitCode={handleSubmitCode}>
-          <ResponseCodeButtons />
+          <ResponseCodeButtons handleAbortButtonOnClick={handleAbortButtonOnClick} />
         </ResponseCodeForm>
       </div>
     </React.Fragment>
