+++ src/components/Signup/SignupEnterCode.tsx
@@ -1,19 +1,15 @@
 import { verifyEmailRequest } from "apis/eduidSignup";
 import EduIDButton from "components/Common/EduIDButton";
+import { ResponseCodeButtons } from "components/Common/ResponseCodeAbortButton";
 import { TimeRemainingWrapper } from "components/Common/TimeRemaining";
 import { ExpiresMeter } from "components/Login/ExpiresMeter";
 import { ResponseCodeForm, ResponseCodeValues } from "components/Login/ResponseCodeForm";
 import { useAppDispatch, useAppSelector } from "eduid-hooks";
 import { Fragment, useContext, useEffect, useState } from "react";
-import { FormRenderProps } from "react-final-form";
 import { FormattedMessage } from "react-intl";
 import { clearNotifications } from "slices/Notifications";
 import { signupSlice } from "slices/Signup";
 import { SignupGlobalStateContext } from "./SignupGlobalState";
-
-interface ResponseCodeButtonsProps {
-  formProps?: FormRenderProps<ResponseCodeValues>;
-}
 
 export function SignupEnterCode(): JSX.Element {
   const signupState = useAppSelector((state) => state.signup.state);
@@ -52,47 +48,6 @@ export function SignupEnterCode(): JSX.Element {
         signupContext.signupService.send({ type: "COMPLETE" });
       }
     }
-  }
-
-  function ResponseCodeButtons(props: ResponseCodeButtonsProps) {
-    if (!props.formProps) {
-      return null;
-    }
-
-    // 'convert' from FormRenderProps to a simple "disabled" boolean
-    return (
-      <ResponseCodeAbortButton
-        disabled={props.formProps.submitting}
-        invalid={props.formProps.invalid}
-        submit={props.formProps.form.submit}
-      />
-    );
-  }
-
-  function ResponseCodeAbortButton(props: { disabled: boolean; invalid: boolean; submit: () => void }) {
-    // abort button usable from both ResponseCodeButtons and when isExpired below
-    return (
-      <div className="buttons">
-        <EduIDButton
-          type="button"
-          buttonstyle="secondary"
-          onClick={handleAbortButtonOnClick}
-          id="response-code-abort-button"
-          disabled={props.disabled}
-        >
-          <FormattedMessage defaultMessage="Cancel" description="Short code form" />
-        </EduIDButton>
-        <EduIDButton
-          type="submit"
-          buttonstyle="primary"
-          onClick={props.submit}
-          id="response-code-ok-button"
-          disabled={props.invalid}
-        >
-          <FormattedMessage defaultMessage="Ok" description="Short code form Ok button" />
-        </EduIDButton>
-      </div>
-    );
   }
 
   if (isExpired) {
@@ -177,7 +132,7 @@ export function SignupEnterCode(): JSX.Element {
 
       <div className="enter-code">
         <ResponseCodeForm inputsDisabled={false} handleSubmitCode={handleSubmitCode}>
-          <ResponseCodeButtons />
+          <ResponseCodeButtons handleAbortButtonOnClick={handleAbortButtonOnClick} />
         </ResponseCodeForm>
       </div>
     </Fragment>
